% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute variance-weighting lambdas from global class proportions}
\usage{
compute_lambda(Y, M)
}
\arguments{
\item{Y}{Integer vector of class labels (1..M).}

\item{M}{Number of classes.}
}
\value{
Numeric vector of length M with lambda_m = 1 / (p_m * (1 - p_m)).
}
\description{
Compute variance-weighting lambdas from global class proportions
}
\keyword{internal}
