% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_effects.R
\name{print.jocf_me}
\alias{print.jocf_me}
\title{Print method for jocf_me objects}
\usage{
\method{print}{jocf_me}(x, digits = 4L, ...)
}
\arguments{
\item{x}{A \code{"jocf_me"} object.}

\item{digits}{Number of significant digits to display. Default \code{4}.}

\item{...}{Currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for jocf_me objects
}
