% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resolve_num_threads}
\alias{resolve_num_threads}
\title{Resolve num.threads to an integer for C++ (0 = all available cores)}
\usage{
resolve_num_threads(num.threads)
}
\description{
Resolve num.threads to an integer for C++ (0 = all available cores)
}
\keyword{internal}
